+++ training_mod_consts/src/files.rs
@@ -1,0 +1,4 @@
+const ULTIMATE_ROOT: &str = "sd:/ultimate/";
+
+const TRAINING_MODPACK_ROOT: &str = &(ULTIMATE_ROOT.to_string() + "TrainingModpack");
+const TRAINING_MODPACK_TOML: &str = &(TRAINING_MODPACK_ROOT.to_string() + "/training_modpack.toml");
